use `other` instead of `self`
